+++ examples/basic/users/rhencloud/meta.nix
@@ -1,0 +1,8 @@
+{
+  hosts = [
+    "nixos-desktop"
+    "hm-standalone"
+  ];
+  uid = 1000;
+  extraGroups = [ "wheel" ];
+}